Create DateValidator - shortest
I checked if the entered date was correct with a simple if function, if it was true then run Case. (As I defined that entered dates have to be in range of 1 - 31, so for the long months I did not need  to check in the entered date was not higher than max date of the month).